+++ navigation/types.ts
@@ -1,0 +1,21 @@
+import type { FormSchema } from '@/components/FormRenderer';
+
+export type TabParamList = {
+  Drafts: undefined;
+  Inbox: undefined;
+  Outbox: undefined;
+  Sent: undefined;
+};
+
+export type RootStackParamList = {
+  Dashboard: undefined;
+  Tabs: { screen?: keyof TabParamList } | undefined;
+  CreateForm: undefined;
+  Form: {
+    schema: FormSchema;
+    formType?: string;
+    formName?: string;
+    data?: Record<string, any>;
+    draftId?: string;
+  };
+};
